UPSTREAM: xskmap: Move non-standard list manipulation to helper
Add a helper in list.h for the non-standard way of clearing a list that is
used in xskmap. This makes it easier to reuse it in the other map types,
and also makes sure this usage is not forgotten in any list refactorings in
the future.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
(cherry picked from commit c8af5cd75e2411d5a5aacf115f59a5ff6b87f3fa)
(cherry picked from commit c8af5cd75e24)